# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CPCMS \u2014 Course Project Configuration and Management System",
    "description": "Internal University Course Project SCM & Configuration Management System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project_api_v1_projects__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Projects"
        ],
        "summary": "Update Project",
        "operationId": "update_project_api_v1_projects__id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/teams/{team_id}/progress": {
      "get": {
        "tags": [
          "Progress"
        ],
        "summary": "List Progress Reports",
        "operationId": "list_progress_reports_api_v1_teams__team_id__progress_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgressReportOut"
                  },
                  "title": "Response List Progress Reports Api V1 Teams  Team Id  Progress Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Progress"
        ],
        "summary": "Submit Progress Report",
        "operationId": "submit_progress_report_api_v1_teams__team_id__progress_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressReportCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/progress/{id}/screenshots": {
      "post": {
        "tags": [
          "Progress"
        ],
        "summary": "Upload Progress Screenshot",
        "operationId": "upload_progress_screenshot_api_v1_progress__id__screenshots_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_progress_screenshot_api_v1_progress__id__screenshots_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/progress/{id}/comment": {
      "post": {
        "tags": [
          "Progress"
        ],
        "summary": "Comment Progress Report",
        "operationId": "comment_progress_report_api_v1_progress__id__comment_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressCommentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_progress_screenshot_api_v1_progress__id__screenshots_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_progress_screenshot_api_v1_progress__id__screenshots_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProgressCommentRequest": {
        "properties": {
          "faculty_comment": {
            "type": "string",
            "minLength": 2,
            "title": "Faculty Comment"
          }
        },
        "type": "object",
        "required": [
          "faculty_comment"
        ],
        "title": "ProgressCommentRequest"
      },
      "ProgressReportCreate": {
        "properties": {
          "review_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Id"
          },
          "seq": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Seq",
            "default": 1
          },
          "requirements_pct": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Requirements Pct"
          },
          "design_pct": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Design Pct"
          },
          "implementation_pct": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Implementation Pct"
          },
          "testing_pct": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Testing Pct"
          },
          "documentation_pct": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Documentation Pct"
          },
          "notes": {
            "type": "string",
            "minLength": 5,
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "requirements_pct",
          "design_pct",
          "implementation_pct",
          "testing_pct",
          "documentation_pct",
          "notes"
        ],
        "title": "ProgressReportCreate"
      },
      "ProgressReportOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team Id"
          },
          "review_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Id"
          },
          "review_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Title"
          },
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "requirements_pct": {
            "type": "integer",
            "title": "Requirements Pct"
          },
          "design_pct": {
            "type": "integer",
            "title": "Design Pct"
          },
          "implementation_pct": {
            "type": "integer",
            "title": "Implementation Pct"
          },
          "testing_pct": {
            "type": "integer",
            "title": "Testing Pct"
          },
          "documentation_pct": {
            "type": "integer",
            "title": "Documentation Pct"
          },
          "overall_progress_pct": {
            "type": "number",
            "title": "Overall Progress Pct",
            "default": 0.0
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "submitted_by": {
            "type": "string",
            "format": "uuid",
            "title": "Submitted By"
          },
          "submitted_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted By Name"
          },
          "faculty_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faculty Comment"
          },
          "screenshots": {
            "items": {
              "$ref": "#/components/schemas/ProgressScreenshotOut"
            },
            "type": "array",
            "title": "Screenshots",
            "default": []
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "team_id",
          "seq",
          "requirements_pct",
          "design_pct",
          "implementation_pct",
          "testing_pct",
          "documentation_pct",
          "notes",
          "submitted_by",
          "created_at"
        ],
        "title": "ProgressReportOut"
      },
...
      "ProjectOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          },
          "course_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Code"
          },
          "course_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Name"
          },
          "faculty_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Faculty User Id"
          },
          "faculty_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faculty Name"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "semester": {
            "type": "string",
            "title": "Semester"
          },
          "academic_year": {
            "type": "string",
            "title": "Academic Year"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "problem_statement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem Statement"
          },
          "objectives": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objectives"
          },
          "min_team_size": {
            "type": "integer",
            "title": "Min Team Size"
          },
          "max_team_size": {
            "type": "integer",
            "title": "Max Team Size"
          },
          "team_formation_mode": {
            "type": "string",
            "title": "Team Formation Mode"
          },
          "team_formation_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Formation Deadline"
          },
          "max_projects_per_student": {
            "type": "integer",
            "title": "Max Projects Per Student"
          },
          "allowed_file_exts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed File Exts"
          },
          "max_file_mb": {
            "type": "integer",
            "title": "Max File Mb"
          },
          "late_policy": {
            "type": "string",
            "title": "Late Policy"
          },
          "late_penalty_percent": {
            "type": "number",
            "title": "Late Penalty Percent"
          },
          "versioning_required": {
            "type": "boolean",
            "title": "Versioning Required"
          },
          "git_required": {
            "type": "boolean",
            "title": "Git Required"
          },
          "baseline_frequency": {
            "type": "string",
            "title": "Baseline Frequency"
          },
          "change_request_required": {
            "type": "boolean",
            "title": "Change Request Required"
          },
          "approval_required": {
            "type": "boolean",
            "title": "Approval Required"
          },
          "branching_policy": {
            "type": "string",
            "title": "Branching Policy"
          },
          "total_marks": {
            "type": "integer",
            "title": "Total Marks"
          },
          "grade_bands": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Grade Bands"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "themes": {
            "items": {
              "$ref": "#/components/schemas/ThemeOut"
            },
            "type": "array",
            "title": "Themes",
            "default": []
          },
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/OutcomeOut"
            },
            "type": "array",
            "title": "Outcomes",
            "default": []
          },
          "ci_templates": {
            "items": {
              "$ref": "#/components/schemas/CITemplateOut"
            },
            "type": "array",
            "title": "Ci Templates",
            "default": []
          },
          "deadlines": {
            "items": {
              "$ref": "#/components/schemas/DeadlineOut"
            },
            "type": "array",
            "title": "Deadlines",
            "default": []
          },
          "evaluation_criteria": {
            "items": {
              "$ref": "#/components/schemas/CriteriaOut"
            },
            "type": "array",
            "title": "Evaluation Criteria",
            "default": []
          },
          "user_team_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Team Id"
          },
          "user_team_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Team Number"
          },
          "user_team_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Team Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_id",
          "faculty_user_id",
          "name",
          "semester",
          "academic_year",
          "min_team_size",
          "max_team_size",
          "team_formation_mode",
          "max_projects_per_student",
          "allowed_file_exts",
          "max_file_mb",
          "late_policy",
          "late_penalty_percent",
          "versioning_required",
          "git_required",
          "baseline_frequency",
          "change_request_required",
          "approval_required",
          "branching_policy",
          "total_marks",
          "grade_bands",
          "status",
          "created_at"
        ],
        "title": "ProjectOut"
      },
...
      "ProjectUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "semester": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Semester"
          },
          "academic_year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Academic Year"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "problem_statement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem Statement"
          },
          "objectives": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objectives"
          },
          "min_team_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Team Size"
          },
          "max_team_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Team Size"
          },
          "team_formation_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Formation Mode"
          },
          "team_formation_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Formation Deadline"
          },
          "max_projects_per_student": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Projects Per Student"
          },
          "allowed_file_exts": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed File Exts"
          },
          "max_file_mb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max File Mb"
          },
          "late_policy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Late Policy"
          },
          "late_penalty_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Late Penalty Percent"
          },
          "versioning_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Versioning Required"
          },
          "git_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Required"
          },
          "baseline_frequency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Frequency"
          },
          "change_request_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Request Required"
          },
          "approval_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Required"
          },
          "branching_policy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branching Policy"
          },
          "total_marks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Marks"
          },
          "grade_bands": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade Bands"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "ProjectUpdate"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}